Fix x64 build: nuint.MaxValue is not a compile-time constant
nuint.MaxValue cannot initialize a const - the width of a native integer is
not known at compile time - so the non-arm64 half of the threshold failed
with CS0133. Only the arm64 side had been compiled locally, so this was
never built.

Type the unbounded case as ulong, which is what MemmoveNativeThreshold above
already does for the same reason. The comparison widens to ulong and still
folds away.

+++ src/libraries/System.Private.CoreLib/src/System/SpanHelpers.ByteMemOps.cs
@@ -45,7 +45,9 @@ namespace System
         // this size it wins. Below it the P/Invoke is still the dominant cost and we come out ahead.
         private const nuint OverlappedForwardThreshold = 1024;
 #else
-        private const nuint OverlappedForwardThreshold = nuint.MaxValue;
+        // ulong rather than nuint because nuint.MaxValue is not a compile-time constant, same as
+        // MemmoveNativeThreshold above. The comparison widens to ulong and folds away.
+        private const ulong OverlappedForwardThreshold = ulong.MaxValue;
 #endif
 #endif
